validation for orders implemented

+++ PerfumeShop.Application/DTO/PerfumeMilliliter/PerfumeMilliliterDto.cs
@@ -5,7 +5,7 @@ namespace PerfumeShop.Application.DTO.PerfumeMilliliter
     public class PerfumeMilliliterDto : BaseDto
     {
         public int PerfumeId { get; set; }
-        public ushort MilliliterId { get; set; }
+        public int MilliliterId { get; set; }
         public decimal UnitPrice { get; set; }
         public int Quantity { get; set; }
     }
